actions/requirements-actions.txt
python-dateutil==2.8.1
numpy~=1.19.2
pytz~=2019.3
ruamel.yaml
langchain==0.0.217
sentence-transformers==2.2.2
sqlalchemy~=1.4
python-dotenv==0.19.2
sendgrid
datasets==2.12.0
faiss-cpu==1.7.4
openai==0.27.6
urllib3==1.26.8
pydantic==1.10.8
typing-extensions==4.4.0
spacy==3.5.4

